extract common/mod.rs
